pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neuronet-ai-cognitiveshell"
version = "0.1.1"
authors = [
  { name="muhammad-robitulloh", email="muhammadrobitulloh19@gmail.com" },
]
description = "An AI-powered assistant for shell commands and code generation, accessible via Telegram and a web dashboard."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Terminals",
]
dependencies = [
    "fastapi~=0.103.2",
    "python-telegram-bot~=21.0.1",
    "requests",
    "python-dotenv~=1.0.1",
    "pexpect~=4.9.0",
    "ansi2html~=1.8.0"
]

[project.urls]
"Homepage" = "https://github.com/muhammad-robitulloh/neuronet-ai-cognitiveshell-web-cli"
"Bug Tracker" = "https://github.com/muhammad-robitulloh/neuronet-ai-cognitiveshell-web-cli/issues"

[project.scripts]
neuronet-ai = "cognitiveshell.cognitiveshell:main"
quickstart = "cognitiveshell.quickstart:main"

[tool.setuptools.packages.find]
include = ["cognitiveshell*"]
